requirements.txt
Django==5.1.1
psycopg2-binary
requests
django-filter==24.3
djangorestframework